requirements.txt
# Dependencias del Pole Bot
discord.py>=2.3.0
python-dotenv>=1.0.0
